# src/main/resources/movie-openapi.yaml
openapi: 3.0.3
info:
  title: API Films
  version: 1.1.0
  description: API Spring Boot de gestion des films et de leur location.

paths:
  /films:
    get:
      summary: Lister les films
      tags: [Films]
      responses:
        '200':
          description: Liste des films
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'
    post:
      summary: Creer un film
      tags: [Films]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '201':
          description: Film cree
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'

  /films/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Recuperer un film par identifiant
      tags: [Films]
      responses:
        '200':
          description: Film trouve
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
        '404':
          description: Film non trouve
    put:
      summary: Mettre a jour un film
      tags: [Films]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MovieInput'
      responses:
        '200':
          description: Film mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'
    delete:
      summary: Supprimer un film
      tags: [Films]
      responses:
        '204':
          description: Film supprime
        '404':
          description: Film non trouve

  /films/genre/{genre}:
    get:
      summary: Rechercher des films par genre
      tags: [Films]
      parameters:
        - name: genre
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Films filtres par genre
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'

  /films/director/{director}:
    get:
      summary: Rechercher des films par realisateur
      tags: [Films]
      parameters:
        - name: director
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Films filtres par realisateur
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'

  /films/artist/{artistId}:
    get:
      summary: Rechercher des films par artiste
      tags: [Films]
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Films filtres par artiste
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Movie'

  /films/{id}/rental:
    patch:
      summary: Ouvrir ou fermer un film a la location et ajuster son prix
      tags: [Location]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RentalUpdate'
      responses:
        '200':
          description: Parametres de location mis a jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Movie'

  /films/import/external:
    post:
      summary: Peupler la base films depuis une source externe
      description: Endpoint prevu pour importer/rafraichir des films via une API externe.
      tags: [Films]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalImportRequest'
      responses:
        '202':
          description: Import lance
        '501':
          description: Non implemente

components:
  schemas:
    ArtistSummary:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string

    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        releaseYear:
          type: integer
        director:
          type: string
        genres:
          type: array
          items:
            type: string
        isOpen:
          type: boolean
        price:
          type: number
          format: double
        minAge:
          type: integer
        artists:
          type: array
          items:
            $ref: '#/components/schemas/ArtistSummary'

    MovieInput:
      type: object
      required: [title, releaseYear, director]
      properties:
        title:
          type: string
        releaseYear:
          type: integer
        director:
          type: string
        genres:
          type: array
          items:
            type: string
        isOpen:
          type: boolean
        price:
          type: number
          format: double
        minAge:
          type: integer

    RentalUpdate:
      type: object
      properties:
        isOpen:
          type: boolean
        price:
          type: number
          format: double

    ExternalImportRequest:
      type: object
      properties:
        source:
          type: string
          example: omdb
        query:
          type: string
          example: science fiction